ProcessCache: Delay removal for custom refcnts
The process cache uses reference counting to ensure we keep hold of
entries that we still need to refer to, but allows us to remove those
that we no longer need. Beyond "process" and "parent", messages could
set custom refcnts on processes. For these we need to keep the process
in the cache until all its descendants have exited (to allow events from
any of the descendants to decrement the reference count on the process).
This is to prevent the stale entry garbage collector from removing the
process while it still has active descendants.

This commit adds a flag on the internal process that specifies that the
process has custom refcnts. It avoids removing the process or its
descendants until all have exited by not decrementing the parent's
refcnt until the process' refcnt is 0.

Signed-off-by: Kevin Sheldrake <kevin.sheldrake@isovalent.com>