enforcer: track missed notifications
This commit adds a enforcer map to track missed
notifications from the enforcer. Notifications are missed when the
enforcer bpf program did not run after the notification happened. This
can, for example, if there is a NotifyEnforcer action from the
raw_syscalls/sys_enter tracepoint without having the enforcer loaded
into this program.

For example:

```
apiVersion: cilium.io/v1alpha1
kind: TracingPolicy
metadata:
  name: "kill-syscalls"
spec:
  enforcers:
  - calls:
    - "sys_swapoff"
    # - "sys_getcpu" <-- mising entry
  tracepoints:
  - subsystem: "raw_syscalls"
    event: "sys_enter"
    args:
    - index: 4
      type: "syscall64"
    selectors:
    - matchArgs:
      - index: 0
        operator: "InMap"
        values:
        - 309 # getcpu
        - 168 # swapoff
      matchActions:
      - action: "NotifyEnforcer"
        argError: -1
        argSig: 9
```

The current version of the code, checks the enforcer map when it updates
it for existing entries. Subsequent patches will improve this.

Executing two getcpu calls from the same thread when the above policy is
loaded, will result in the following entries:

```
bpftool map dump  pinned /sys/fs/bpf/tetragon/kill-syscalls/enforcer_missed_notifications
[{
        "key": {
            "act_info": {
                "func_id": 0,
                "arg": 309
            },
            "reason": 1
        },
        "value": 1
    }
]
```

Note that the syscall id is recorded so that we can know what was the
missed syscall.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>