enforcer: add CleanupEnforcerNotification action
Previous commits introduced a bpf map to track missed enforcer
notifications, by checking the enforcer map for existing entries when
attempting to set a new one.

As an improvement, we want to introduce an action allowing to check for
enforcer missed notifications. The intention is to use this action when
exiting system calls.

This commit introduces the policy and proto changes.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>